Added a separate webSocketProvider and wrapped it around the app
Closes #102, closes #103, closes #104
