+++ app/layout.tsx
@@ -4,6 +4,7 @@ import { App, ConfigProvider, theme } from "antd";
 import { AntdRegistry } from "@ant-design/nextjs-registry";
 import "@/styles/globals.css";
 //import "antd/dist/reset.css"; // For the confirmation prompt (handleDeletion()) on exoplanet profile page
+import WebSocketProvider from "@/components/WebSocketProvider";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -67,8 +68,9 @@ export default function RootLayout({
           }}
         >
           <App>
-
-          <AntdRegistry>{children}</AntdRegistry>
+            <WebSocketProvider>            
+              <AntdRegistry>{children}</AntdRegistry>
+            </WebSocketProvider>
           </App>
           
         </ConfigProvider>
